pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "sm_bluesky"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
description = "Bluesky code for Diamond's surface and magnetic materials beamline."
dependencies = [
    #
    # If a constraint is not set here or if the dependency is pinned to a hash
    # it will be auto-pinned to the latest release version by the pre-release workflow
    #
    "bluesky",
    #"my_bluesky @ file:///scratch/bluesky_development/my_bluesky",
    "dls-dodal@git+https://github.com/DiamondLightSource/dodal.git@single_electron_analyser_detector",
    "ophyd-async[sim]",
    "scanspec",
]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "copier",
    "myst-parser",
    "pre-commit",
    "pydata-sphinx-theme>=0.12",
    "pyright",
    "pytest",
    "pytest-cov",
    "ruff",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinx-design",
    "tox-uv",
    "types-mock",
    "pytest-asyncio",

    #usefull stuff for testing
    "jupyter",
    "ipython",
    "matplotlib",
    "numpy",
    "pyepics",
    "blueapi",

]

[project.scripts]
sm-bluesky = "sm_bluesky.__main__:main"

[project.urls]
GitHub = "https://github.com/DiamondLightSource/sm-bluesky"

[[project.authors]] # Further authors may be added by duplicating this section
email = "raymond.fan@diamond.ac.uk"
name = "Raymond Fan"

[tool.setuptools_scm]
version_file = "src/sm_bluesky/_version.py"

[tool.pyright]
typeCheckingMode = "standard"
reportMissingImports = false  # Ignore missing stubs in imported modules

[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
addopts = """
    --tb=native -vv --doctest-modules --doctest-glob="*.rst"
    """
# https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
filterwarnings = "error"
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"
asyncio_mode = "auto"


[tool.coverage.run]
patch = ["subprocess"]
data_file = "/tmp/sm_bluesky.coverage"

[tool.coverage.paths]
# Tests are run from installed location, map back to the src directory
source = ["src", "**/site-packages/"]

[tool.tox]
skipsdist = true
# envs to runs automatically with tox -p
env_list = ["pre-commit", "type-checking", "tests", "docs"]

[tool.tox.env_run_base]
# Pass all environment variables to the test environments
passenv = ["*"]

[tool.tox.env.pre-commit]
description = "Run pre-commit on all files"
commands = [
    [
        "pre-commit",
        "run",
        "--all-files",
        "--show-diff-on-failure",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.type-checking]
description = "Run pyright"
commands = [
    [
        "pyright",
        "--pythonpath",
        "{env:VIRTUAL_ENV}/bin/python",
        "src",
        "tests",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.tests]
description = "Run tests with coverage"
commands = [
    [
        "pytest",
        "--cov=sm_bluesky",
        "--cov-report",
        "term",
        "--cov-report",
        "xml:cov.xml",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.docs]
description = "Run docs build with clean environment"
commands = [
    [
        "sphinx-build",
        "--fresh-env",
        "--show-traceback",
        "--fail-on-warning",
        "--keep-going",
        "docs",
        "build/html",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.docs-autobuild]
description = "Run docs with autobuild on change"
commands = [
    [
        "sphinx-autobuild",
        "--show-traceback",
        "--watch",
        "README.md",
        "docs",
        "build/html",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.ruff]
src = ["src", "tests"]
line-length = 88
lint.select = [
    "B",   # flake8-bugbear - https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "C4",  # flake8-comprehensions - https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "E",   # pycodestyle errors - https://docs.astral.sh/ruff/rules/#error-e
    "F",   # pyflakes rules - https://docs.astral.sh/ruff/rules/#pyflakes-f
    "N",   # pep8-naming - https://docs.astral.sh/ruff/rules/#pep8-naming-n
    "W",   # pycodestyle warnings - https://docs.astral.sh/ruff/rules/#warning-w
    "I",   # isort - https://docs.astral.sh/ruff/rules/#isort-i
    "UP",  # pyupgrade - https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "SLF", # self - https://docs.astral.sh/ruff/settings/#lintflake8-self
]
lint.extend-ignore = [
    "B008", # Allow inject("device")
]

[tool.ruff.lint.flake8-import-conventions.extend-aliases]
# We often shorten "Annotated" to "A" for brevity
"typing.Annotated" = "A"


[tool.ruff.lint.per-file-ignores]
# By default, private member access is allowed in tests
# See https://github.com/DiamondLightSource/python-copier-template/issues/154
# Remove this line to forbid private member access in tests
"tests/**/*" = ["SLF001"]
